Fix warnings

+++ src/TortoiseProc/ProgressCommands/ResetProgressCommand.cpp
@@ -54,8 +54,8 @@ bool ResetProgressCommand::Run(CGitProgressList* list, CString& sWindowTitle, in
 	checkout_options.progress_cb = [](const char* path, size_t completed_steps, size_t total_steps, void* payload)
 	{
 		CGitProgressList::Payload* cbpayload = (CGitProgressList::Payload*)payload;
-		cbpayload->list->m_itemCountTotal = total_steps;
-		cbpayload->list->m_itemCount = completed_steps;
+		cbpayload->list->m_itemCountTotal = (int)total_steps;
+		cbpayload->list->m_itemCount = (int)completed_steps;
 		cbpayload->list->AddNotify(new CGitProgressList::WC_File_NotificationData(CUnicodeUtils::GetUnicode(path), CGitProgressList::WC_File_NotificationData::git_wc_notify_checkout));
 	};
 	if (git_reset(repo, target, (git_reset_t)(m_resetType + 1), &checkout_options, nullptr, nullptr))
